refactor(persistence): ADR-0035 — split audit_chain pure core into kirra-audit-hash (enabling slice 2)
The crux kirra-persistence enabling slice. verifier_store's ~30 references to
crate::audit_chain were the biggest coupling blocking the extraction; the pure
half is now a lean leaf crate.

New crate kirra-audit-hash (deps: sha2, hex, ed25519-dalek — no rusqlite, no
state) holds the PURE audit/causal chain primitives, moved VERBATIM:
- compute_record_hash_v1 / compute_record_hash / compute_record_hash_v2
- compute_causal_record_hash + CausalRecordHashInput
- canonical_signing_payload{,_v2} / canonical_anchor_head_payload /
  canonical_causal_signing_payload / canonical_causal_anchor_head_payload
- verifying_key_id

Root src/audit_chain.rs keeps the STATEFUL AuditChainLinker (append-into-a-tx,
rusqlite); its compute_record_hash* associated fns now delegate to the crate, and
it re-exports the crate so every crate::audit_chain::<fn> path (service bin,
tests, other root modules) resolves unchanged. verifier_store's verify/read path
(audit.rs, mod.rs, fabric.rs) now names kirra_audit_hash::* directly, removing
that coupling.

Byte-exactness is load-bearing (these encoders ARE the on-disk audit/causal-log
format). Verified byte-identical: power-loss audit drill (audit_chain_prefix_on_kill,
3 — reopens the WAL and re-verifies the chain), two_node_rollout EP-13
audit-anchored (4), `cargo test --lib` 760 (incl. audit-chain verify +
break_audit_chain tamper tests), fmt, guardrails, orphan-cores. Detached
lockfiles (fuzz / kirra-verifier-pg / parko) refreshed for the new crate.

Remaining before the wholesale verifier_store -> kirra-persistence move: slice 2b
(invert the ChainedAuditAppender write seam — the AuditAppender trait moves down,
the impl injected from root) + the small couplings (ShippedAuditRecord,
is_valid_verdict_id, KeyRegistry).
